share core declarations
