+++ packages/bippy/src/index.ts
@@ -1,4 +1,4 @@
 import "./install-hook-only.js";
 
-export * from "./core.js";
+export * from "./core-implementation.js";
 export * from "./react.js";
